+++ e2e/modal-navigation-ng/e2e/modal-layout.e2e-spec.ts
@@ -1,10 +1,10 @@
 import { AppiumDriver, createDriver } from "nativescript-dev-appium";
-import { Screen } from "./screen"
+import { Screen } from "./screens/screen"
 import {
     roots,
     testNestedModalPageBackground,
     testDialogBackground,
-} from "./shared.e2e-spec"
+} from "./screens/shared-screen"
 
 describe("modal-layout:", () => {